backend/app/core/token_utils.py
from jose import JWTError, jwt

from app.core.config import get_settings


def decode_subject(token: str) -> str | None:
    settings = get_settings()
    try:
        payload = jwt.decode(token, settings.secret_key, algorithms=[settings.jwt_algorithm])
    except JWTError:
        return None
    return payload.get("sub")

