backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scopeguard-ai-backend"
version = "0.1.0"
description = "Authorization-gated AI security assessment platform"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "celery>=5.4.0",
  "docker>=7.1.0",
  "dnspython>=2.6.1",
  "email-validator>=2.1.1",
  "fastapi>=0.111.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.4",
  "neo4j>=5.21.0",
  "passlib[bcrypt]>=1.7.4",
  "pgvector>=0.2.5",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "psycopg[binary]>=3.1.19",
  "pydantic-settings>=2.3.4",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.7",
  "reportlab>=4.2.2",
  "slowapi>=0.1.9",
  "sqlalchemy>=2.0.31",
  "strawberry-graphql[fastapi]>=0.229.0",
  "tenacity>=8.4.2",
  "typer>=0.12.3",
  "uvicorn[standard]>=0.30.1"
]

[project.scripts]
scopeguard-cli = "app.cli:main"

[tool.setuptools.packages.find]
include = ["app*"]

